VME_IRQ_Controller: made IRQ line level-sensitive.
There are two reasons for doing so:
- compatibility with Wishbone and the VIC interrupt controller
- possibility of losing an edge-triggered IRQ and hanging
interrupts when different cores trigger interrupts very close to each
other. The modified interrupter implements a retry
mechanism, that is, if the IRQ line gets stuck for longer than
certain period (g_retry_timeout), an IRQ cycle is repeated on the VME bus.

Signed off Tomas